+++ src/components/CreateAccount/CreateAccount.jsx
@@ -5,7 +5,13 @@ import styled from "styled-components";
 const Wrapper = styled.div`
   padding: 1rem;
 `;
+
+const auth = new GoTrue({
+  APIUrl: "https://eazy-meals.netlify.app/.netlify/identity",
+});
+
 const CreateAccount = () => {
+
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [message, setMessage] = useState(null);
@@ -13,8 +19,30 @@ const CreateAccount = () => {
   const handlePassword = (event) => setPassword(event.target.value);
   const handleSubmit = async (event) => {
     console.log("success", email, password);
+    const fakeToken = {
+      access_token: "example-jwt-token",
+      token_type: "bearer",
+      expires_in: 3600,
+      refresh_token: "example-refresh_token",
+      expires_at: 1526062471000,
+    };
+    window.sessionStorage.setItem("token", JSON.stringify(fakeToken));
+    const retrievedToken = window.sessionStorage.getItem("token");
+    fetch("/example", {
+      headers: { Authorization: `Bearer ${retrievedToken.access_token}` },
+    });
+    try {
+      setMessage("Logging in...");
+      event.preventDefault();
+      const response = await auth.login(email, password);
+      if (response.description) setMessage(response.description);
+      if (response.token) {
+        window.sessionStorage("token", JSON.stringify(response.token));
+      }
+    } catch (error) {
+      setMessage(error.toString());
+    }
   };
-
   return (
     <div>
       <div>
@@ -43,10 +71,10 @@ const CreateAccount = () => {
           </Wrapper>
 
           <button type="submit" onClick={handleSubmit}>
-            Sign In
+            Create Account
           </button>
           <button type="submit" onClick={handleSubmit}>
-            Create Account
+            Cancel
           </button>
         </form>
       </div>
