+++ src/components/LoginAccount/LoginAccount.jsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import GoTrue from "gotrue-js";
-import Faker from "Faker";
 
 const UserName = (props) => {
   const [user, setUser] = useState(false);
@@ -19,12 +18,18 @@ const UserName = (props) => {
   }
 };
 
-const auth = new GoTrue({
-  APIUrl: "https://eazy-meals.netlify.app/.netlify/identity",
-});
 const Wrapper = styled.div`
   padding: 1rem;
 `;
+
+//  const auth = new goTrue({
+//   auth.login(email, password, remember)
+//    .login(email.value, password.value, true)
+//    .then((response) => {
+//      showMessage(`Success! Response: ${JSON.stringify({ response })}`, form);
+//   })
+//   .catch((error) => showMessage(`Failed :( ${JSON.stringify(error)}`, form));
+// });
 const LoginAccount = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
@@ -33,34 +38,12 @@ const LoginAccount = () => {
   const handlePassword = (event) => setPassword(event.target.value);
   const handleSubmit = async (event) => {
     console.log("success", email, password);
-    const fakeToken = {
-      access_token: "example-jwt-token",
-      token_type: "bearer",
-      expires_in: 3600,
-      refresh_token: "example-refresh_token",
-      expires_at: 1526062471000,
-    };
-    window.sessionStorage.setItem("token", JSON.stringify(fakeToken));
-    const retrievedToken = window.sessionStorage.getItem("token");
-    fetch("/example", {
-      headers: { Authorization: `Bearer ${retrievedToken.access_token}` },
-    });
-    try {
-      setMessage("Logging in...");
-      event.preventDefault();
-      const response = await auth.login(email, password);
-      if (response.description) setMessage(response.description);
-      if (response.token) {
-        window.sessionStorage("token", JSON.stringify(response.token));
-      }
-    } catch (error) {
-      setMessage(error.toString());
-    }
   };
+
   return (
     <div>
       <div>
-        <h1>Create Account</h1>
+        <h1>Login Account</h1>
         {message && <div>{message}</div>}
         <form onSubmit={handleSubmit}>
           <Wrapper>
@@ -84,9 +67,7 @@ const LoginAccount = () => {
             />
           </Wrapper>
           <Wrapper type="submit" required onClick={handleSubmit}>
-          <button >
-            Submit
-          </button>
+            <button>Submit</button>
           </Wrapper>
         </form>
       </div>
